Added rake task subscribe_recent_adopters.
This task should be run periodically to subscribe recent adopters to an email list managed
by GovDelivery. It subscribes all users for whom the new subscribed_at column is null.
The column is updated after they are subscribed. To resubscribe a user, that column must be
set back to null.